Fix to #7290 - Query compilation error or data corruption for queries with GroupJoin wrapped in a subquery when the subset of columns is being projected
Problem was that we were incorrectly reasoning about which query sources need to be materialized in the GroupJoin scenario. Previously, when we encountered GroupJoin, we marked its main from clause and the joined clause for materialization, but we would not do that recursively. This causes problems if the GroupJoin (e.g. optional navigation) is in a subquery - we would not mark it for materialization. This would lead to compilation errors in some cases, and in others (when there was a client-evaluation component) to data corruption.

Naive implementation of the fix - marking every query source of the GroupJoin for materialization would result in performance regression (we would start to materialize more unnecessary columns - in many scenarios they are not needed). Therefore we also implemented an enhancement to the logic that chooses whether query source (only MainFromClause for now) needs to be materialized.

We don't need to materialize MainFromClause for the following case:

(
  (if we encounter GroupJoinClause followed by AdditionalFromClause, which references the group join and contains subquery with a DefaultIfEmpty - basically GroupJoin-SelectMany-DefaultIfEmpty pattern that can be flattened to LOJ and client GroupJoin will be removed)
  AND
  (reference to the group is NOT present anywhere else in the query model - if the group is referenced then we need to preserve client-side GroupJoin operation, which requires MainFromClause to be materialized)
)
AND
(
  (reference to MainFromClause is NOT projected in the selector - we need to materialize because it might be needed in the parent query model)
  OR
  (queryModel is wrapped in a result operator that returns a scalar - e.g. Count, even if we project MainFromClause, the wrapping result operator makes it inaccessible from the parent query, meaning that it's not needed after all)
)

Second part of the fix is that in LiftSubquery, when we update query sources, we would sometimes try to rewrite EntityShapers into ValueBufferShapers (if the lifted query itself is not marked for materialization). This however is not safe to do in case of GroupJoin. Fix is to only perform this conversion shapers inside ShapedQuery method.